Improve Missing Value / `null` Handling (#49)
XGBoostSharp currently uses `-1.0F` as a constant to represent missing
values. This is contrary to the official Python library, which uses
`float.NaN`. See, e.g., the parameter missing
[here](https://github.com/mdabros/XGBoostSharp/pull/DMatrix).
XGBoostSharp currently does not allow its users to change that value.

Changes in this PR:

- Allow users to change the constant that represents a missing value,
similar to the Python and C API.
- Change the default missing value constant to float.NaN. While
technically this is a breaking change, it can probably be considered a
bug fix as well.
- Add a new constructor that automatically handles input and feature
arrays containing null values.

Note: I have not yet E2E tested the change and found it hard to set up a
meaningful unit test. I wanted to open the PR ASAP, though, to get
@mdabros's opinion on this.